backend/app/services/face_service.py
# Note: face_recognition requires complex dependencies (CMake, dlib)
# This is a mock implementation for development purposes
import numpy as np
from PIL import Image
import io
from typing import Tuple, Optional
import random

class FaceService:
    def __init__(self):
        self.tolerance = 0.6  # Default tolerance for face matching
    
    def load_image_from_bytes(self, image_bytes: bytes) -> np.ndarray:
        """Load image from bytes and convert to RGB numpy array"""
        try:
            pil_image = Image.open(io.BytesIO(image_bytes))
            # Convert to RGB if needed
            if pil_image.mode != 'RGB':
                pil_image = pil_image.convert('RGB')
            return np.array(pil_image)
        except Exception as e:
            raise Exception(f"Error loading image: {str(e)}")
    
    def extract_face_encoding(self, image_bytes: bytes) -> Optional[np.ndarray]:
        """Extract face encoding from image (mock implementation)"""
        try:
            # Load image
            image = self.load_image_from_bytes(image_bytes)
            
            # Mock face detection - just return a random encoding for demo
            # In production, this would use actual face detection
            print(f"Mock: Processing image of size {image.shape}")
            
            # Create a mock face encoding (128-dimensional vector)
            mock_encoding = np.random.random(128).astype(np.float64)
            
            return mock_encoding
        
        except Exception as e:
            raise Exception(f"Error extracting face encoding: {str(e)}")
    
    def compare_faces(self, aadhaar_image_bytes: bytes, live_image_bytes: bytes, tolerance: float = None) -> Tuple[bool, float]:
        """Compare faces from Aadhaar photo and live photo (mock implementation)"""
        try:
            if tolerance is None:
                tolerance = self.tolerance
            
            # Extract face encodings from both images
            aadhaar_encoding = self.extract_face_encoding(aadhaar_image_bytes)
            live_encoding = self.extract_face_encoding(live_image_bytes)
            
            if aadhaar_encoding is None or live_encoding is None:
                raise Exception("Could not extract face encodings from one or both images")
            
            # Mock face comparison - simulate realistic results
            # In production, this would use actual face recognition algorithms
            face_distance = random.uniform(0.2, 0.8)  # Simulate distance
            
            # Convert distance to confidence percentage
            confidence = max(0, (1 - face_distance) * 100)
            
            # Check if faces match
            is_match = face_distance <= tolerance
            
            print(f"Mock: Face comparison - distance: {face_distance:.3f}, confidence: {confidence:.1f}%, match: {is_match}")
            
            return is_match, confidence
        
        except Exception as e:
            raise Exception(f"Error comparing faces: {str(e)}")
    
    def validate_face_quality(self, image_bytes: bytes) -> dict:
        """Validate face quality in image (mock implementation)"""
        try:
            image = self.load_image_from_bytes(image_bytes)
            
            # Mock face detection - simulate finding one face
            # In production, this would use actual face detection
            
            validation_result = {
                'is_valid': True,
                'issues': [],
                'face_count': 1  # Mock: assume one face is found
            }
            
            # Basic image quality checks
            image_height, image_width = image.shape[:2]
            
            if image_width < 100 or image_height < 100:
                validation_result['is_valid'] = False
                validation_result['issues'].append('Image resolution is too low')
            
            print(f"Mock: Face validation for image {image_width}x{image_height} - Valid: {validation_result['is_valid']}")
            
            return validation_result
        
        except Exception as e:
            raise Exception(f"Error validating face quality: {str(e)}")

# Global face service instance
face_service = None

def get_face_service() -> FaceService:
    """Get face service instance"""
    global face_service
    if face_service is None:
        face_service = FaceService()
    return face_service