backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow>=10.0.0
opencv-python==4.8.1.78
easyocr>=1.7.2
face-recognition==1.3.0
numpy==1.24.3
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
aiofiles==23.2.1